requirements.txt
# Core
fastapi==0.109.2
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart

# Monitoring
sentry-sdk==1.44.0
prometheus-fastapi-instrumentator==7.1.0

# AI/ML
langchain-core>=0.2.20
langchain>=0.2.0
langchain-community>=0.2.0
langchain_openai>=0.1.7
transformers>=4.39.0
sentence-transformers>=2.6.0
pypdf==4.2.0
faiss-cpu==1.8.0
unstructured==0.13.7

# Utils
numpy==1.26.4
pandas==2.2.1
requests==2.31.0

# UI
streamlit==1.32.0
